backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.1

# AI / LLM / RAG
langchain==0.3.14
langchain-openai==0.3.12
langchain-community==0.3.14
faiss-cpu==1.13.2

# File Processing
pypdf==5.1.0
python-docx==1.1.2

# Audio/Video Transcription
openai>=1.68.2

# Object Storage (MinIO via S3 API)
boto3==1.36.2

# Background Tasks
celery[redis]==5.4.0
redis==5.2.1

# Auth
python-jose[cryptography]==3.3.0
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.25.0

# Utilities
python-dotenv==1.0.1
uuid6==2024.7.10
